CV
Basic CV using HTML and CSS This is a website which gives a brief description of my passion and my skills. I have used basic and advanced HTML and CSS with the goal of getting used to them and for practice. I have given links to my Twitter handle, Github account and LinkedIn profile. One can also send me an email using the Contact Me button on the website.